move js to directory
